refactor: consolidate AssemblyTestScenario fixture into conftest.py and update related tests
